requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Testing
pytest>=7.4.0

# Translation (NEW - Hugging Face API)
python-dotenv>=1.0.0

# Old translation library (REPLACED - keep for rollback)
# argostranslate>=1.9.0  # Replaced by Hugging Face Qwen2.5-7B-Instruct
